requirements.txt
kiteconnect==4.2.0
python-dotenv==1.0.0
websocket-client==1.6.4
cryptography==41.0.7
pydantic==2.5.0
structlog==23.2.0

# Notification dependencies
twilio==8.10.0
aiohttp==3.9.1
requests==2.31.0

# Automated token generation dependencies
selenium==4.15.2
webdriver-manager==4.0.1
pyotp==2.9.0
